Move integrations test to infrastructure layer.
